Split OperatorSpacingSniff, use upstream sniffs
