+++ src/Cdn77CodingStandard/ruleset.xml
@@ -20,10 +20,8 @@
         <exclude name="SlevomatCodingStandard.ControlStructures.ControlStructureSpacing"/>
         <!-- replaced by SlevomatCodingStandard.ControlStructures.BlockControlStructureSpacing and ...JumpStatementsSpacing -->
         <exclude name="Squiz.WhiteSpace.ControlStructureSpacing"/>
-        <!-- replaced by Cdn77CodingStandard.OperatorSpacingSniff -->
-        <exclude name="Generic.WhiteSpace.IncrementDecrementSpacing"/>
-        <!-- replaced by Cdn77CodingStandard.OperatorSpacingSniff -->
-        <exclude name="Squiz.WhiteSpace.LogicalOperatorSpacing"/>
+        <!-- replaced by Cdn77CodingStandard.Operators.OperatorSpacing -->
+        <exclude name="Squiz.WhiteSpace.OperatorSpacing"/>
     </rule>
 
     <!-- compatibility with PSR-12 and JetBrains defaults -->
@@ -68,5 +66,11 @@
     <!-- Ignore blank lines before if, foreach, etc @todo replace with Slevomat sniffs once merged into upstream -->
     <rule ref="Cdn77CodingStandard.ControlStructures.BlockControlStructureSpacing">
         <exclude name="Cdn77CodingStandard.ControlStructures.BlockControlStructureSpacing.IncorrectLinesCountBeforeControlStructure"/>
+    </rule>
+
+    <rule ref="Cdn77CodingStandard.Operators.OperatorSpacing">
+        <properties>
+            <property name="allowMultipleStatementsAlignment" value="false"/>
+        </properties>
     </rule>
 </ruleset>
