+++ tests/Sniffs/Operators/data/NegationOperatorSpacingSniffNoErrors.php
@@ -1,0 +1,10 @@
+<?php
+
+$a = $a-$b;
+$a = $a  -  $b;
+$a =  -$a;
+$a =  self::CONST-  $a;
+$a = -(15 + 3);
+if ($a > -16 || $a > -$b || $a > -self::CONSTANT) {
+
+}
