+++ tests/Sniffs/Operators/data/OperatorSpacingSniffErrors.fixed.php
@@ -30,24 +30,13 @@ $a = $a * $b;
 $a = $a + $b;
 $a = $a ^ $b;
 
-$a = ++$a;
-$a = --$this->a;
-$a = -$a;
 $a = self::CONST - $a;
 
-$a = $a++;
-$a = $this->a--;
-$a = (++$a);
-$a = $a[++$a];
-$a = $a[$a++];
-
 $a += $b;
 $a -= $b;
 
 $a += $b;
 $a -= $b;
 
-$a = $a . $b;
-$a = $a . $b;
 $a = ($a) instanceof $b;
 $a = $a instanceof $b;
