+++ tests/Sniffs/Operators/data/OperatorSpacingSniffErrors.php
@@ -30,16 +30,7 @@ $a = $a  *  $b;
 $a = $a  +  $b;
 $a = $a  ^  $b;
 
-$a =  ++ $a;
-$a =  -- $this->a;
-$a =  - $a;
 $a =  self::CONST-  $a;
-
-$a =  $a ++;
-$a =  $this->a --;
-$a = (++ $a);
-$a = $a[++ $a];
-$a = $a[$a ++];
 
 $a+=$b;
 $a-=$b;
@@ -47,7 +38,5 @@ $a-=$b;
 $a  +=  $b;
 $a  -=  $b;
 
-$a = $a.$b;
-$a = $a  .  $b;
 $a = ($a)instanceof$b;
 $a = $a  instanceof  $b;
