+++ tests/Sniffs/Operators/data/OperatorSpacingSniffNoErrors.php
@@ -15,14 +15,12 @@ $a = $a * $b;
 $a = $a + $b;
 $a = $a ^ $b;
 
-++$a;
---$a;
-
 $a += $b;
 $a -= $b;
 
-$a = $a . $b;
 $a = $a instanceof $b;
 $a = (-$b);
 $a = $a +
     $b;
+
+$a = $a ?: true;
